+++ assets/components/card-content.component.jsx
@@ -9,12 +9,14 @@ cardContent.mixins = [
 ];
 
 cardContent.propTypes = {
-  className: React.PropTypes.string
+  className: React.PropTypes.string,
+  renderArrow: React.PropTypes.bool
 };
 
 cardContent.getDefaultProps = function cardContentGetDefaultProps() {
   return {
-    className: null
+    className: null,
+    renderArrow: false
   };
 };
 
@@ -28,9 +30,25 @@ cardContent.getCssClasses = function cardContentGetCssClasses() {
   return cssClasses;
 };
 
+cardContent.renderArrow = function cardContentRenderArrow() {
+  var arrow = null;
+
+  if (this.props.renderArrow === true) {
+    arrow = (
+      <span className="card__content-arrow">
+        <div className={'card__content-arrow-border m-' + this.props.arrowPosition}></div>
+        <div className={'card__content-arrow-background m-' + this.props.arrowPosition}></div>
+      </span>
+    );
+  }
+
+  return arrow;
+};
+
 cardContent.render = function cardContentRender() {
   return (
     <div className={this.getCssClasses().join(' ')}>
+      {this.renderArrow()}
       {this.props.children}
     </div>
   );