+++ assets/components/card-header.component.jsx
@@ -10,18 +10,22 @@ cardHeader.mixins = [
 
 cardHeader.propTypes = {
   className: React.PropTypes.string,
-  renderArrow: React.PropTypes.bool
+  renderArrow: React.PropTypes.bool,
+  arrowPosition: React.PropTypes.oneOf(['top', 'bottom'])
 };
 
 cardHeader.getDefaultProps = function cardHeaderGetDefaultProps() {
   return {
     className: null,
-    renderArrow: false
+    renderArrow: false,
+    arrowPosition: 'bottom'
   };
 };
 
 cardHeader.getCssClasses = function cardHeaderGetCssClasses() {
   var cssClasses = ['card__header'];
+
+  cssClasses.push();
 
   if (this.props.className) {
     cssClasses = cssClasses.concat(this.props.className.split(' '));
@@ -33,9 +37,9 @@ cardHeader.getCssClasses = function cardHeaderGetCssClasses() {
 cardHeader.renderArrow = function cardHeaderRenderArrow() {
   var arrow = null;
 
-  if(this.props.renderArrow === true) {
+  if (this.props.renderArrow === true) {
     arrow = (
-      <div className="card__header-arrow"></div>
+      <div className={'card__header-arrow m-' + this.props.arrowPosition}></div>
     );
   }