+++ test/specs/assets/components/card-header.component.spec.jsx
@@ -20,7 +20,7 @@ describe('card header component', function() {
     expect(cardHeaderArrow.length).to.equal(0);
   });
 
-  it('should render arrow', function() {
+  it('should be able to render arrow', function() {
     this.component = React.render(<CardHeader renderArrow={true}>header</CardHeader>, div);
     var cardHeader = reactTestUtils.findRenderedDOMComponentWithClass(this.component, 'card__header');
     var cardHeaderArrow = reactTestUtils.scryRenderedDOMComponentsWithClass(this.component, 'card__header-arrow');
@@ -28,6 +28,14 @@ describe('card header component', function() {
     expect(cardHeader.props.className).to.equal('card__header');
     expect(cardHeader.props.children[0]).to.equal('header');
     expect(cardHeaderArrow.length).to.equal(1);
+    expect(cardHeaderArrow[0].props.className).to.equal('card__header-arrow m-bottom');
+  });
+
+  it('should be able to render arrow on the top', function() {
+    this.component = React.render(<CardHeader renderArrow={true} arrowPosition="top">header</CardHeader>, div);
+    var cardHeaderArrow = reactTestUtils.scryRenderedDOMComponentsWithClass(this.component, 'card__header-arrow');
+
+    expect(cardHeaderArrow[0].props.className).to.equal('card__header-arrow m-top');
   });
 
   it('should be able to add custom classes', function() {